Logout and toggling tasks work.

+++ src/main/java/com/example/logintodo/repositories/PersonRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface PersonRepository extends CrudRepository<Person,Integer> {
+public interface PersonRepository extends CrudRepository<Person,Long> {
 
     Optional<Person> findByUserName(String userName);
     Optional<Person> findByEmail(String Email);
